mv loadLayers to context
